refactor(ai-chat): fold ResponseStopped into SystemMessage
- git mv SystemMessage.{tsx,scss} to components/
- add responseStopped prop; reads messages_responseStopped from store
- absorb ResponseStopped.scss rules into SystemMessage.scss
- swap <ResponseStopped /> for <SystemMessage responseStopped />
- re-point SystemMessage imports to components/ in all call sites
- delete ResponseStopped.{tsx,scss}
- clean up stale @use entries and comment

Closes #1432
