--- packages/ai-chat/src/chat/components-legacy/ResponseStopped.tsx
@@ -1,29 +1,0 @@
-/*
- *  Copyright IBM Corp. 2025, 2026
- *
- *  This source code is licensed under the Apache-2.0 license found in the
- *  LICENSE file in the root directory of this source tree.
- *
- *  @license
- */
-
-import React from "react";
-
-import { useSelector } from "../hooks/useSelector";
-import { AppState } from "../../types/state/AppState";
-
-/**
- * This component is for displaying a message under streamed responses that were stopped.
- */
-function ResponseStopped() {
-  const messages_responseStopped = useSelector(
-    (state: AppState) => state.languagePack.messages_responseStopped,
-  );
-  return (
-    <div className="cds-aichat--response-stopped">
-      {messages_responseStopped}
-    </div>
-  );
-}
-
-export { ResponseStopped };
