+++ packages/ai-chat/src/chat/components-legacy/_imports.scss
@@ -10,8 +10,6 @@
 @use "input/imports" as input;
 @use "launcher/imports" as launcher;
 @use "responseTypes/imports" as responseTypes;
-@use "ResponseStopped" as ResponseStopped;
-@use "SystemMessage" as SystemMessage;
 @use "MessageComponent";
 @use "MessagesComponent";
 @use "@carbon/styles/scss/_carbon-utilities.scss" as carousel;
