+++ packages/ai-chat/src/chat/components/SystemMessage.scss
@@ -59,3 +59,28 @@
   margin-block-start: layout.$spacing-07;
   padding-block-start: layout.$spacing-05;
 }
+
+// Response stopped indicator (absorbed from ResponseStopped.scss)
+.cds-aichat--response-stopped {
+  @include type.type-style("body-01");
+
+  color: theme.$text-secondary;
+  font-style: italic;
+  margin-block-start: layout.$spacing-05;
+  text-align: center;
+}
+
+.cds-aichat--standard-width
+  .cds-aichat--conversational-search
+  + .cds-aichat--response-stopped,
+.cds-aichat--wide-width
+  .cds-aichat--conversational-search
+  + .cds-aichat--response-stopped {
+  padding-inline-start: 56px;
+}
+
+.cds-aichat--narrow-width
+  .cds-aichat--conversational-search
+  + .cds-aichat--response-stopped {
+  padding-inline-start: 16px;
+}
