+++ packages/ai-chat/src/chat/components/_imports.scss
@@ -11,3 +11,4 @@
 @use "util/imports" as utils;
 @use "homeScreen/imports" as homescreen;
 @use "util/TextBlock/TextBlock.scss" as textblock;
+@use "SystemMessage";
